♻️ :: PersistenceAdapter 구조 변경
